+++ UnitTests/Geometry/EPATests.cpp
@@ -53,17 +53,17 @@ TEST_SUITE("EPATests")
 			// Check angle between v1 and v2
 			float angle = AngleBetweenVectors(v1, v2);
 			CHECK(angle < 0.1f);
-			EPA_TESTS_TRACE("Angle = %.9g\n", angle);
+			EPA_TESTS_TRACE("Angle = %.9g\n", (double)angle);
 
 			// Check delta between contact on A
 			Vec3 dpa = pa2 - pa1;
 			CHECK(dpa.Length() < 8.0e-4f);
-			EPA_TESTS_TRACE("Delta A = %.9g\n", dpa.Length());
+			EPA_TESTS_TRACE("Delta A = %.9g\n", (double)dpa.Length());
 
 			// Check delta between contact on B
 			Vec3 dpb = pb2 - pb1;
 			CHECK(dpb.Length() < 8.0e-4f);
-			EPA_TESTS_TRACE("Delta B = %.9g\n", dpb.Length());
+			EPA_TESTS_TRACE("Delta B = %.9g\n", (double)dpb.Length());
 		}
 
 		return intersect1;
